Added Severity Levels for Intervention

+++ lib/shared/globals.dart
@@ -4,28 +4,28 @@ bool emailVerificationEnabled = true;
 int maxScale = 7; // what will be considered high or True in 0-10 range sliders
 
 // beats per minute (Heart Rate)
-double minPossibleHeartRate = 20;   // Near-death, rare survival
-double minExtremeHeartRate = 30;    // Severe bradycardia, medical emergency
-double minNormalHeartRate = 60;     // Typical lower bound of normal resting
-double maxNormalHeartRate = 100;    // Upper bound of normal resting
-double maxExtremeHeartRate = 220;   // Near maximum for exertion (220 - age)
-double maxPossibleHeartRate = 300;  // Fatal arrhythmias occur
+int minPossibleHeartRate = 20;   // Near-death, rare survival
+int minExtremeHeartRate = 30;    // Severe bradycardia, medical emergency
+int minNormalHeartRate = 60;     // Typical lower bound of normal resting
+int maxNormalHeartRate = 100;    // Upper bound of normal resting
+int maxExtremeHeartRate = 220;   // Near maximum for exertion (220 - age)
+int maxPossibleHeartRate = 300;  // Fatal arrhythmias occur
 
 // mmHg (Blood Pressure - Systolic)
-double minPossibleBloodPressureSystolic = 40.0;   // Circulatory collapse
-double minExtremeBloodPressureSystolic = 60.0;   // Severe hypotension (shock)
-double minNormalBloodPressureSystolic = 90.0;    // Lower bound of normal
-double maxNormalBloodPressureSystolic = 130.0;   // High-normal upper bound
-double maxExtremeBloodPressureSystolic = 180.0;  // Hypertensive emergency
-double maxPossibleBloodPressureSystolic = 300.0; // Lethal hypertensive crisis
+int minPossibleBloodPressureSystolic = 40;   // Circulatory collapse
+int minExtremeBloodPressureSystolic = 60;   // Severe hypotension (shock)
+int minNormalBloodPressureSystolic = 90;    // Lower bound of normal
+int maxNormalBloodPressureSystolic = 130;   // High-normal upper bound
+int maxExtremeBloodPressureSystolic = 180;  // Hypertensive emergency
+int maxPossibleBloodPressureSystolic = 300; // Lethal hypertensive crisis
 
 // mmHg (Blood Pressure - Diastolic)
-double minPossibleBloodPressureDiastolic = 20.0;  // Severe shock, organ failure
-double minExtremeBloodPressureDiastolic = 40.0;  // Dangerously low
-double minNormalBloodPressureDiastolic = 60.0;   // Lower bound of normal
-double maxNormalBloodPressureDiastolic = 80.0;   // Healthy upper bound
-double maxExtremeBloodPressureDiastolic = 120.0; // Hypertensive emergency
-double maxPossibleBloodPressureDiastolic = 200.0;// Near-fatal pressure
+int minPossibleBloodPressureDiastolic = 20;  // Severe shock, organ failure
+int minExtremeBloodPressureDiastolic = 40;  // Dangerously low
+int minNormalBloodPressureDiastolic = 60;   // Lower bound of normal
+int maxNormalBloodPressureDiastolic = 80;   // Healthy upper bound
+int maxExtremeBloodPressureDiastolic = 120; // Hypertensive emergency
+int maxPossibleBloodPressureDiastolic = 200;// Near-fatal pressure
 
 // percentage (Blood Oxygen Saturation - SpO₂)
 double minPossibleOxygenSaturation = 30.0;  // Severe hypoxia, near death
